+++ pages/party/[category].tsx
@@ -1,12 +1,20 @@
+import { useMemo } from "react";
+
 import type { NextPage } from "next";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { CATEGORY } from "src/constants";
 
 import { Layout } from "@components/common";
+import { CategoryBar } from "@components/party";
 
 const PartyList: NextPage = () => {
   const router = useRouter();
+
+  const headerConfig = useMemo(() => {
+    return { bottomBorder: false };
+  }, []);
+
   return (
     <>
       <Head>
@@ -14,7 +22,9 @@ const PartyList: NextPage = () => {
           {CATEGORY[router.query.category as keyof typeof CATEGORY]} 파티 목록
         </title>
       </Head>
-      <Layout isShowBottomNavigation={false}>파티 목록</Layout>
+      <Layout headerConfig={headerConfig} isShowBottomNavigation={false}>
+        <CategoryBar />
+      </Layout>
     </>
   );
 };
