+++ src/components/common/Layout/BottomNavigation/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./BottomNavigationButton";
