+++ src/components/common/Layout/BottomNavigation/interfaces.ts
@@ -1,0 +1,5 @@
+export interface NavigationInfo {
+  icon: "home" | "chatBubble" | "user";
+  link: "/" | "/chat/list" | "/my-page";
+  title: "홈" | "채팅" | "My";
+}
