--- src/components/home/Categories/interfaces.ts
@@ -1,7 +1,0 @@
-import { CategoryKey, CategoryName } from "@types";
-
-export interface Category {
-  image: string;
-  name: CategoryName;
-  path: CategoryKey;
-}
