+++ src/styles/globals.css
@@ -1,10 +1,13 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
 html,
 body {
   font-size: 62.5%;
   max-width: 45rem;
   min-height: 100vh;
-  padding: 0;
-  margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
@@ -43,11 +46,6 @@ a {
   text-decoration: none;
   color: white;
   cursor: pointer;
-}
-
-ol, ul {
-  margin: 0;
-  padding: 0;
 }
 
 li {
